refactor(canvas): extract utilities constants into constants.ts
